% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crse.R
\name{clustSE}
\alias{clustSE}
\title{Cluster robust standard errors with degrees of freedom adjustments}
\usage{
clustSE(mod, clust, data = NULL, digits = 4, ztest = FALSE)
}
\arguments{
\item{mod}{The \code{lm} model object.}

\item{clust}{The cluster variable (with quotes).}

\item{digits}{Number of decimal places to display.}

\item{ztest}{If a normal approximation should be used as the naive degrees of freedom. If FALSE, the HLM degrees of freedom will be used.}
}
\value{
All the CR adjustments with p-values.
}
\description{
Function to compute the CR0, CR1, CR2 cluster
robust standard errors with Bell and McCaffrey (2002)
degrees of freedom adjustments. Useful when dealing with datasets with a few clusters.
}
\examples{
clustSE(lm(mpg ~ am + wt, data = mtcars), 'cyl')

}
\references{
\cite{Bell, R., & McCaffrey, D. (2002). Bias reduction in standard errors for linear regression with multi-stage samples. Survey Methodology, 28, 169-182.
(\href{https://www150.statcan.gc.ca/n1/pub/12-001-x/2002002/article/9058-eng.pdf}{link})}
}
