% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr2_mixed.R
\name{cr2_mixed}
\alias{cr2_mixed}
\title{Cluster robust standard errors with degrees of freedom adjustments}
\usage{
cr2_mixed(m1, digits = 4, satt = FALSE)
}
\arguments{
\item{m1}{The \code{merMod} model object.}

\item{digits}{Number of decimal places to display.}

\item{satt}{If Satterthwaite degrees of freedom are to be computed}
}
\value{
A data frame with the CR adjustments with p-values.
\item{estimate}{The regression coefficient.}
\item{se.unadj}{The model-based (regular, unadjusted) SE.}
}
\description{
Function to compute the CR2 cluster
robust standard errors (SE) with Bell and McCaffrey (2002)
degrees of freedom (dof) adjustments.
}
\details{
EXPERIMENTAL
}
\references{
\cite{Bell, R., & McCaffrey, D. (2002). Bias reduction in standard errors for linear regression with multi-stage samples. Survey Methodology, 28, 169-182.
(\href{https://www150.statcan.gc.ca/n1/pub/12-001-x/2002002/article/9058-eng.pdf}{link})}

Liang, K.Y., & Zeger, S. L. (1986). Longitudinal data analysis using generalized linear models. \emph{Biometrika, 73}(1), 13–22.
\doi{10.1093/biomet/73.1.13}
}
