% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getV.R
\name{getV}
\alias{getV}
\title{Get V matrix for merMod objects}
\usage{
getV(x)
}
\arguments{
\item{x}{lme4 object}
}
\value{
V matrix (weight) for multilevel models
}
\description{
Function to extract V matrix.
}
