% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{MatSqrtInverse}
\alias{MatSqrtInverse}
\title{Compute the inverse square root of a matrix}
\usage{
MatSqrtInverse(A)
}
\arguments{
\item{A}{The matrix object.}
}
\description{
From Imbens and Kolesar (2016).
}
