% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedat.R
\docType{data}
\name{sharedat}
\alias{sharedat}
\title{Project SHARE}
\format{
A data frame with 5399 observations and 7 variables.
}
\source{
\href{https://doi.org/10.7910/DVN/YXMQZM}{Harvard dataverse}
}
\usage{
data(sharedat)
}
\arguments{
\item{school}{The cluster variable.}

\item{sex}{factor indicating F or M}

\item{arm}{treatment arm = 1 vs control = 0}

\item{kscore}{Pupil knowledge of sexual health}

\item{idno}{student id number}

\item{sc}{factor showing the highest social class of the father or mother
based on occupation (coded 10: I (highest), 20: II,
31: III non-manual, 32: III manual, 40: IV, 50: V (lowest), 99: not coded).}

\item{zscore}{standardized knowledge score}
}
\description{
Project SHARE (Sexual Health and Relationships) was a cluster randomized trial (CRT) in
Scotland carried out to measure the impact of a school-based
sexual health program (Wight et al., 2002).
}
\examples{
data(sharedat)
}
\references{
Moulton, L. (2015). readme.txt contains an overall explanation of the data sets \link{Data set}. Harvard.
(\href{https://doi.org/10.7910/DVN/YXMQZM}{link})

Wight, D., Raab, G. M., Henderson, M., Abraham, C., Buston, K., Hart, G., & Scott, S. (2002). Limits of teacher delivered sex education:
Interim behavioural outcomes from randomised trial. BMJ, 324, 1430.
(\href{https://doi.org/10.1136/bmj.324.7351.1430}{doi})
}
\keyword{datasets}
