% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crse.R
\name{clustSE}
\alias{clustSE}
\title{Cluster robust standard errors with degrees of freedom adjustments}
\usage{
clustSE(mod, clust, data = NULL, digits = 4, ztest = FALSE)
}
\arguments{
\item{mod}{The \code{lm} model object.}

\item{clust}{The cluster variable (with quotes).}

\item{digits}{Number of decimal places to display.}

\item{ztest}{If a normal approximation should be used as the naive degrees of freedom. If FALSE, the HLM degrees of freedom will be used.}
}
\value{
All the CR adjustments with p values
}
\description{
Function to compute the CR0, CR1, CR2 cluster
robust standard errors with Bell and McCaffrey (2002)
degrees of freedom adjustments. Useful when dealing with datasets with a few clusters.
}
\examples{
data(mtcars)
ex1 <- lm(mpg ~ wt + am, data = mtcars)
clustSE(ex1, 'cyl')




}
